+++ mainapp/static/already_seen_movies.txt
@@ -1,0 +1,14 @@
+I have already seen below movies, do not recommend these:
+
+Harry Potter and the Deathly Hallows: Part 2 2011
+Harry Potter and the Goblet of Fire 2005
+Harry Potter and the Chamber of Secrets 2002
+Harry Potter and the Sorcerer's Stone 2001
+The Matrix Revolutions 2003
+The Matrix 1999
+The Miracle of the Bells 1948
+Eastern Promises 2007
+The Dark Knight 2008
+Frost/Nixon 2008
+The Tillman Story 2010
+Birth 2004
