+++ mainapp/static/watchlist.txt
@@ -1,0 +1,4 @@
+Recommend me only movies from below watchlist:
+
+Okja 2017
+Star Wars: Episode IX - The Rise of Skywalker 2019
